Moved chart styled-components to ChartComponent/styles.js
