+++ src/components/ChartComponent/styles.js
