--- src/components/styled-components/__tests__/ChartContainer.test.js
@@ -1,7 +1,0 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { ChartContainer } from "../ChartContainer";
-
-test("render ChartContainer", () => {
-  shallow(<ChartContainer />);
-});
